refactor: extract editor mixins
