+++ mainappsrc/version.py
@@ -1,5 +1,5 @@
 """Project version information."""
 
-VERSION = "0.2.31"
+VERSION = "0.2.32"
 
 __all__ = ["VERSION"]
